email_assistant/requirements.txt
djangorestframework==3.12.4
pymongo==3.11.4
dnspython==1.16.0

qdrant-client>=1.7
openai>=1.12
cohere>=5.0
PyJWT>=2.8
python-dotenv>=1.0
cryptography>=42.0
tiktoken>=0.6
imapclient>=3.0
pydantic>=2.5

Django==3.1.12
djongo==1.3.7
sqlparse==0.2.4
django-cors-headers==3.7.0

PyMuPDF>=1.24
python-docx>=1.1
openpyxl>=3.1
